server/requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
django-nested-admin==3.1.3
django-templated-mail==1.1.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
python-monkey-business==1.0.0
pytz==2018.9
six==1.12.0
